separate model and db logic for users
