+++ lib/models/task.js
@@ -107,7 +107,6 @@ class Task {
      * Fetches all tasks in scope
      */
     static async list(userId) {
-        // TODO: separate model and db logic
         const items = await schemas.tasks.listTasksForUser(userId);
         return items.map(item => TaskData.fromJSON(item));
     }
@@ -118,7 +117,6 @@ class Task {
      * @param {String} taskId - task id
      */
     static async getById(userId, taskId) {
-        // TODO: separate model and db logic
         const item = await schemas.tasks.getTaskById(userId, taskId);
         return TaskData.fromJSON(item);
     }
