+++ lib/models/user.js
@@ -4,7 +4,6 @@
 
 const uuid = require('uuid');
 const schemas = require('../db/schemas');
-const dynamodb = require('../db/dynamodb_client');
 
 /**
  * Module
@@ -22,33 +21,8 @@ class UserData {
         return new UserData(uuid.v4(), name, nowISO);
     }
 
-    static fromDynamoDb(data) {
-        const expectedKeys = ['id', 'name', 'created'];
-        expectedKeys.forEach((key) => {
-            if (!Object.keys(data).includes(key)) {
-                throw new Error(`Missing key for user: ${key}`);
-            }
-        });
-
-        if (Object.keys(data).length != expectedKeys.length) {
-            throw new Error(`Expected user data to have keys ${expectedKeys}, got ${Object.keys(data)}`);
-        }
-
-        return new UserData(data.id.S, data.name.S, data.created.S);
-    }
-
-    toDynamoDB() {
-        return {
-            id: {
-                S: this.id
-            },
-            name: {
-                S: this.name
-            },
-            created: {
-                S: this.created
-            },
-        };
+    static fromJSON(data) {
+        return new UserData(data.id, data.name, data.created);
     }
 
     toJSON() {
@@ -64,11 +38,7 @@ class User {
     static async create(data) {
         // TODO: validate userId in authorization
         const userData = UserData.create(data.name);
-        const payload = userData.toDynamoDB();
-        await dynamodb.putItem({
-            TableName: 'Users',
-            Item: payload
-        }).promise();
+        await schemas.users.create(userData.toJSON());
         // TODO: handle errors
         return userData;
     }
@@ -78,21 +48,8 @@ class User {
      * @param {String} userId - user id
      */
     static async getById(userId) {
-        const response = await dynamodb.getItem({
-            TableName: 'Users',
-            Key: {
-                id: {
-                    S: userId
-                }
-            }
-        }).promise();
-        if (!response.Item) {
-            const error = new Error(`User with id ${userId} not found`);
-            error.httpStatusCode = 404;
-            throw error;
-        }
-        const user = UserData.fromDynamoDb(response.Item);
-        return user;
+        const data = await schemas.users.getById(userId);
+        return UserData.fromJSON(data);
     }
 
     static async removeAll() {
