+++ lib/routes/user.js
@@ -24,7 +24,7 @@ async function handleGetUser(req, res, next) {
     try {
         const userData = await models.user.User.getById(userId);
         logResponse(requestId, 200, req);
-        res.json(userData);
+        res.json(userData.toJSON());
     } catch (error) {
         logger.error(`Error getting user: ${error}`);
         error.requestId = requestId;
